Create new deploy subpackage
No functional changes, just breaking out the deployer.py
into separate modules.
