+++ chalice/utils.py
@@ -1,0 +1,39 @@
+import os
+
+
+class OSUtils(object):
+    def open(self, filename, mode):
+        # type: (str, str) -> IO
+        return open(filename, mode)
+
+    def remove_file(self, filename):
+        # type: (str) -> None
+        """Remove a file, noop if file does not exist."""
+        # Unlike os.remove, if the file does not exist,
+        # then this method does nothing.
+        try:
+            os.remove(filename)
+        except OSError:
+            pass
+
+    def file_exists(self, filename):
+        # type: (str) -> bool
+        return os.path.isfile(filename)
+
+    def get_file_contents(self, filename, binary=True):
+        # type: (str, bool) -> str
+        if binary:
+            mode = 'rb'
+        else:
+            mode = 'r'
+        with open(filename, mode) as f:
+            return f.read()
+
+    def set_file_contents(self, filename, contents, binary=True):
+        # type: (str, str, bool) -> None
+        if binary:
+            mode = 'wb'
+        else:
+            mode = 'w'
+        with open(filename, mode) as f:
+            f.write(contents)
