+++ tests/integration/test_features.py
@@ -1,13 +1,14 @@
-import pytest
-import os
-import requests
 import json
-import botocore.session
+import os
 import shutil
 
-from chalice import deployer
+import botocore.session
+import pytest
+import requests
+
 from chalice.cli import load_chalice_app
 from chalice.config import Config
+from chalice.deploy import deployer
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
 PROJECT_DIR = os.path.join(CURRENT_DIR, 'testapp')
